app:: APHudScene: delete
